Update Company table

+++ JobBoard.Core/Feutures/Companies/Commands/Handler/CompanyCommandHandler.cs
@@ -1,10 +1,14 @@
-﻿using AutoMapper;
+﻿using System.Security.Claims;
+using AutoMapper;
 using JobBoard.Core.Bases;
 using JobBoard.Core.Feutures.Companies.Commands.Models;
 using JobBoard.Core.Resources;
+using JobBoard.Core.Security.Requirements;
 using JobBoard.Data.Entities;
 using JobBoard.Service.Abstractions;
+using JobBoard.Service.Authentication.Interfaces;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Localization;
 
 namespace JobBoard.Core.Feutures.Companies.Commands.Handler
@@ -16,14 +20,25 @@ namespace JobBoard.Core.Feutures.Companies.Commands.Handler
 		#region Fields 
 		private readonly ICompanyService _companyService;
 		private readonly IMapper _mapper;
+		private readonly IStringLocalizer<SharedResources> _stringLocalizer;
+		private readonly IAuthorizationService _authorizationService;
+		private readonly ICurrentUserService _currentUserService;
 
 		#endregion
 
 		#region constructors
-		public CompanyCommandHandler(ICompanyService companyService, IMapper mapper, IStringLocalizer<SharedResources> stringLocalizer) : base(stringLocalizer)
+		public CompanyCommandHandler(ICompanyService companyService,
+									IMapper mapper,
+									IStringLocalizer<SharedResources> stringLocalizer,
+									IAuthorizationService authorizationService,
+									ICurrentUserService currentUserService
+									) : base(stringLocalizer)
 		{
 			_companyService = companyService;
 			_mapper = mapper;
+			_stringLocalizer = stringLocalizer;
+			_authorizationService = authorizationService;
+			_currentUserService = currentUserService;
 		}
 		#endregion
 
@@ -41,8 +56,12 @@ namespace JobBoard.Core.Feutures.Companies.Commands.Handler
 		// update handler
 		public async Task<Response<int>> Handle(UpdateCompanyCommand request, CancellationToken cancellationToken)
 		{
-			//bool Exist = await _companyService.IsExistByIdAsync(request.CompanyId);
+
 			var company = await _companyService.GetCompanyByIdAsync(request.CompanyId);
+
+			var isAuthorized = await _authorizationService.AuthorizeAsync(new ClaimsPrincipal(), company, new CompanyOwnerRequirement());
+
+			if (!isAuthorized.Succeeded) return Forbidden<int>(_stringLocalizer[SharedResourcesKeys.NoAccess]);
 
 			if (company == null) return NotFound<int>("There is no Company to Ipdate! Make sure to enter the correct Id");
 
@@ -56,6 +75,17 @@ namespace JobBoard.Core.Feutures.Companies.Commands.Handler
 		public async Task<Response<string>> Handle(DeleteCompanyCommand request, CancellationToken cancellationToken)
 		{
 			var company = await _companyService.GetCompanyByIdAsync(request.Id);
+
+			var userRoles = await _currentUserService.GetCurrentUserRoles();
+
+			if (!userRoles.Contains("Admin"))
+			{
+				var isAuthorized = await _authorizationService.AuthorizeAsync(new ClaimsPrincipal(), company, new CompanyOwnerRequirement());
+
+				if (!isAuthorized.Succeeded) return Forbidden<string>(_stringLocalizer[SharedResourcesKeys.NoAccess]);
+
+			}
+
 			if (company == null) return BadRequest<string>($"There Is No Country With Id = {request.Id}");
 
 			await _companyService.DeleteAsync(company);
